Split `Between` condition to `Between` and `NotBetween`
