+++ src/QueryBuilder/AbstractDQLQueryBuilder.php
@@ -519,7 +519,7 @@ abstract class AbstractDQLQueryBuilder implements DQLQueryBuilderInterface
             'OR' => Condition\OrX::class,
             '=' => Condition\Equals::class,
             'BETWEEN' => Condition\Between::class,
-            'NOT BETWEEN' => Condition\Between::class,
+            'NOT BETWEEN' => Condition\NotBetween::class,
             'IN' => Condition\In::class,
             'NOT IN' => Condition\In::class,
             'LIKE' => Condition\Like::class,
@@ -552,6 +552,7 @@ abstract class AbstractDQLQueryBuilder implements DQLQueryBuilderInterface
             Condition\AndX::class => Condition\Builder\LogicalBuilder::class,
             Condition\OrX::class => Condition\Builder\LogicalBuilder::class,
             Condition\Between::class => Condition\Builder\BetweenBuilder::class,
+            Condition\NotBetween::class => Condition\Builder\BetweenBuilder::class,
             Condition\In::class => Condition\Builder\InBuilder::class,
             Condition\Like::class => Condition\Builder\LikeBuilder::class,
             Condition\Equals::class => Condition\Builder\EqualsBuilder::class,
